CODE REHAUL Fixed save characters
Info:
Characters now save to a file when they are first created,
when the user clicks Choose on the char select page,
or when the Save button is tapped on the stats page.
If a single character is selected it is no longer set to currentCharacter

TODO:
Make times in format xx:xx:xx
Fix landscape mode? LinearLayout gets squished in landscape